Refactoring : Extract class refactoring adding FileOperationUtil class to maintain solid principle
