+++ biojava-structure/src/test/java/org/biojava/nbio/structure/align/util/AtomCacheTest.java
@@ -39,8 +39,7 @@ import java.util.Date;
 import java.util.List;
 import java.util.Locale;
 import java.util.zip.GZIPOutputStream;
-
-import org.biojava.nbio.core.util.FileDownloadUtils;
+import org.biojava.nbio.core.util.FileOperationUtils;
 import org.biojava.nbio.structure.AtomPositionMap;
 import org.biojava.nbio.structure.Chain;
 import org.biojava.nbio.structure.Group;
@@ -370,7 +369,7 @@ public class AtomCacheTest {
 		System.setProperty(UserConfiguration.PDB_DIR, tmpCache.toString());
 		System.setProperty(UserConfiguration.PDB_CACHE_DIR, tmpCache.toString());
 
-		FileDownloadUtils.deleteDirectory(tmpCache);
+		FileOperationUtils.deleteDirectory(tmpCache);
 		Files.createDirectory(tmpCache);
 		try {
 			cache.setPath(tmpCache.toString());
@@ -390,7 +389,7 @@ public class AtomCacheTest {
 			Files.createDirectories(testCif.getParent());
 			URL resource = AtomCacheTest.class.getResource("/atp.cif.gz");
 			File src = new File(resource.getPath());
-			FileDownloadUtils.copy(src, testCif.toFile());
+			FileOperationUtils.copy(src, testCif.toFile());
 
 			// Load structure
 			Structure s = cache.getStructure("1ABC");
@@ -410,7 +409,7 @@ public class AtomCacheTest {
 			assertTrue(chem.getAtoms().size() > 0);
 			assertEquals("NON-POLYMER", chem.getType());
 		} finally {
-			FileDownloadUtils.deleteDirectory(tmpCache);
+			FileOperationUtils.deleteDirectory(tmpCache);
 		}
 	}
 
@@ -429,7 +428,7 @@ public class AtomCacheTest {
 		System.setProperty(UserConfiguration.PDB_DIR, tmpCache.toString());
 		System.setProperty(UserConfiguration.PDB_CACHE_DIR, tmpCache.toString());
 
-		FileDownloadUtils.deleteDirectory(tmpCache);
+		FileOperationUtils.deleteDirectory(tmpCache);
 		Files.createDirectory(tmpCache);
 		try {
 			cache.setPath(tmpCache.toString());
@@ -453,7 +452,7 @@ public class AtomCacheTest {
 			Files.createDirectories(testCif.getParent());
 			URL resource = AtomCacheTest.class.getResource("/atp.cif.gz");
 			File src = new File(resource.getPath());
-			FileDownloadUtils.copy(src, testCif.toFile());
+			FileOperationUtils.copy(src, testCif.toFile());
 
 			// Load structure
 			Structure s = cache.getStructure("1ABC");
@@ -473,7 +472,7 @@ public class AtomCacheTest {
 			assertTrue(chem.getAtoms().size() > 0);
 			assertEquals("NON-POLYMER", chem.getType());
 		} finally {
-			FileDownloadUtils.deleteDirectory(tmpCache);
+			FileOperationUtils.deleteDirectory(tmpCache);
 		}
 	}
